requirements.txt
Django
djangorestframework
psycopg2
python-dateutil
sqlalchemy